more baseline hugo config; shuffle into dirs.
The amount of dirs that get spammed at the root of a project with
hugo in its default configuration just doesn't feel nice.

I'd re-seat the "content" dir right to the root if it was possible,
but it's tetchy: you can't give it patterns for what's content nor
what to ignore, so it'll pick up the .git dir and... so on.
I've seen someone effectively re-home their entire site with symlinks,
but... nah.  Fine.  We'll have a content dir.