Bulk import existing docs content; enable menu.
The default menu in this theme is... pretty nice; I'm still expecting
to use its override features later, though.  Turns out a good ToC is
one of the single most important things to navigable docs, and so I'm
starting to have quite some opinions on the details of it.

There will be some issues with the new theme; this content was written
against a different one (and I'm switching theme at the same time as
starting with this clean repo).  In this commit, the theme isn't
even enabled on any of this content, because it actually requires
opt-in rather than all being "_default" like the old one was -- that's
a bit one.  Oh, and notices are missing correct css; that's smaller.

Gonna handle stuff in a long series of commits because making progress
otherwise is too hard.  Many will probably have rendering artifacts.