turn the default layout into book theme docs.
with a symlink, cause, ya know.

remove the other layout files: they were copies of the ones from the
theme, in two different directory names... because I was initially
thinking about re-orienting all docs to come in those two top level
directories.  I've since rethought that (didn't seem like the thing
to do right now; may consider again later, but it was just too much),
which is why the last couple commits have been fairly broken.

The other alternative is to tag literally every page with a header
of `type = "docs"`.  And I'd... rather not.

Somehow this seems to be making the menu contain a bunch of things
twice