massive warning: hugo versions
I'm starting to get more and more fun details of hugo versions fucking
with me in various ways.

I've been working with hugo-extended-v0.55.6 for a while, and it's
*mostly* been doing the right thing... but now in toggling back and
forth between that and hugo-extended-v0.61.0, they're both wrong:

- v0.61 reports some things differently that makes this version of
 this theme double-list some pages in the TOC; cool cool...

- v0.55 doesn't let me actually do the layout/_default symlink and
 emits a fairly inane error about it being a directory...? ok...

- v0.55 also is making some pages report 404 with this theme,
 and others are just blank, varying on if they're "/_index.md"-style
 or if they're a dir full of singles.  (I had confusing as heck
 behavior around this distinction in the other theme I was using too,
 so it's good to know this is endemic to hugo ecosystem rather than
 being just one theme; super thrilled.)

I've been trying really, really hard to work with this tool, because
in theory, it describes itself as doing what it wants; and in theory,
it should be on the minimalist side of sysadmin headaches.

... but those claims and those hopes really aren't holding up.

I'm going to switch fully to 0.61 and also yank a bunch of updates
to this theme up next, and hopefully that just fixes everything; but
it remains true that the amount of coupling of versions is staggering,
and the degree of error messages that come out when something is wrong
is nearly none; and this combination does not bring reliability or joy.