some mild alternative preference on spacing.
mostly to prove I can.  As always, the appropriate filename to use
in order to hit the magic spot in hugo's logic took guessing.

I don't remember all the other things I tried, because this is a
cleanup commit of experiments started last summer and then left in
a pile for a few months, but of all the other files laying around,
this appears to be the one that actually _did something_; the rest
are being ditched.

(There was a "css.css" file in this dir that pastself opted not to use?
I'm not sure why.  Maybe it's because putting stuff in this scss file
causes it to get compiled into the css already used by the rest of the
theme headers?  And adding any other css files added more epicycles?
Probably that.  idk.)