Use redirect at / to /welcome.  Hugo bugs?  In my index?  It's more likely than you think.
Seriously, what is with this stuff.  It's like... literally the first impression.

Despite there being a placeholder in the theme saying
"yeah, put a homepage at 'content/_index.md'", that
ABSOLUTELY
DOES NOT WORK.

I tried it with this theme against versions 0.55 and 0.61 of hugo
(which as remarked in previous commits, was enough to a change to
varyingly break and fix several things) and in neither does it do
a gol'dangged thing to put a file there.

I had different problems around this file in the other theme, too,
though admittedly different ones; the 'content/_index.md' page in
the bulk import was a dupe of one of the main content pages for
*reasons* (intense handwaving).

Why is the home page so special and so hard in hugo?
You like... can't render it... until you understand everything
about themes and layouts and page types and single-vs-list and...
just... what.  whathwathwat.  And even then, it's apparently
such an entangled mess that even theme authors still regularly
do something somewhere in such a way that the home page gets
mishandled or otherwise breaks behavior you'd expect for _any other_
incidence of '_index.md'?  Sure.  Super good.  Totally fine.

Putting a meta redirect in the 'home.html' file cuts through this.

And yes, that html file *does* override the theme's placeholder page.

Unlike 'content/_index.md'.

Why is this file called "home" when it's got html in it,
when "_index" is the magic prefix for markdown files,
and neither of those is the "index.html" that the last SEVERAL DECADES
of web convention dictates?

I dunno.

Because hugo.

My eyes literally could not roll harder without escaping orbit.

I hope after another few timeunits of configuring this and papering
over the most absurd of issues and bugs, I reach a plateau where the
sunk costs are worth it again.  Because right now, this is infuriating.
I'm seriously looking at alternatives to hugo written in PHP and
feeling just about wistful enough to try them.