remove workflows draft page.
It turns out the 'draft' system in hugo -- I'm really on a roll for
finding bugs and design flaws in hugo, tonight, aren't I? -- doesn't
really work particularly reliably.  It's better and safer to just
not have draft files.

In this theme, it generates text with the page title in the TOC menu.
It doesn't link anywhere; it just generates... dead text.  Neato.

I'm not sure if this a bug of the theme or of the core tool, but again,
this is a problem I experienced -- in slightly different ways! -- with
BOTH this theme and the previous one I'm in process of migrating from.
So I think there's a pretty strong argument to be made here that even
if the bug is *technically* in the theme... it's a kind of bug that's
endemic to the hugo *ecosystem*, which indicates a design flaw in hugo.