oh cool hugo 0.60 processes markdwon differently
I mostly agree with the security logic in https://github.com/gohugoio/hugo/issues/6581

but also with the "wow that might've deserved a verbose warning" as
suggested in some Discourse threads on the subject.

It's a heck of a breaking change, IMO.  Most markdown in most places
is widely understood to be something that can contain HTML.
Even github allows (most?) html by default.
One has to give up and use tables sometimes.
This isn't a terribly unusual corner case.

Ah well.  Moving on.