New draft on why-aka-vision page.
A decent amount of this comes from another previous draft here:
https://gist.github.com/warpfork/d7f8c9ca622f9c7e29538d32989365ed

I go back and forth on how much belongs on one page versus what
should be shifted to sub-pages, because a good amount of philosophic
perspective is required to make it clear what makes this project
distinctive and worth it, but brevity is also valuable...
Maybe this round will hit the balance.

The strategy components at the end are still a todo, clearly.
But fragments of what goes there just needs to be collected from
other docs and reprocessed for local format.  (Interestingly,
I'm actually backing a bit away from giving "hermetic" a large
amount of airtime.  It's important, but also, a detail -- it's
arguably necessary, but certainly not *sufficient* to reach the
overall desired goals, which makes it a poor disambiguator for
why this project as a whole is distinctive and important.
And of course, also (quite happily!) it's becoming a fairly
"obvious", "table-stakes" sort of thing in the development world
at large, which I'm frankly downright excited to take in stride.)